Gnawing on parameters for particle properties.
I feel like there should be a nice way to generalize from the rand()
crate so that we can not only generate random values of any type
(color, vector, etc), but can generate different distributions of them.
It appears to be something involving the rand::distributions::Sample
trait and maybe the RandSample struct and something or other...

But that gets Tricky and I don't fully understand how it works, and
honestly all we're ever really going to want is a uniform or gaussian
distribution for numbers, vectors or colors, so for now I'm just
going to hard-code it a bit.