Integrating Scene with the GameState trait is trickier than it looks.
Things like the "load" function get a little screwy, so it's probably
best to not try to mess with that in isolation until I understand the
problem better.

The thing is...  Scene's and the SceneManager don't map *exactly* to
the GameState trait, though they're conceptually very close.  But it's
kind of a GameState+extra stuff, which doesn't behave quite the same,
and which the GameState trait itself doesn't encompass.