Struggling with scene manager.
I think that I'm trying to build TOO MUCH dependency injection.
This is leading to lots of awkward dependency loops where things
within the SceneManager have to touch things outside of it and
initialization is all topsy-turvy and such.  Mainly 'cause we
already do dependency injection in the GameState trait in ggez,
which bascially has a load() callback to do some initialization
code without having any knowledge of what it is...  Having the
SceneManager forward that to an initial scene is a good idea,
but then the initial scene is going to need to load a bunch of
stuff and add more scenes dynamically, which ain't possible right now,
which gets silly fast.

Though upon consideration having ALL loading taking place in the
load() callback makes it impossible to do things like splash screens,
so that's not the right place to actually initialize crap anyway,
it seems.  I'll fiddle around with that a bit more, but if it continues
being a pain then I'll just make SceneManager not implement GameState
after all and get on with life the brute force way.