Stubbing out dealing with multiple component types.
It's a little tricky 'cause we suddenly have to deal with the
possibility of sparseness, which we really just don't do right now.
Should probably just use Option but that feels a bit wonky.
Though not as wonky as making there be uninitialized values everywhere.
Need to do some research.